src/main/java/SortArray.java
import java.util.Arrays;

public class SortArray {

    public int[] sortArray(int[] array) {
        if (array.length >= 2) {
            int n = array.length;
            int temp = 0;
            for (int i = 0; i < n; i++) {
                for (int j = 1; j < (n - i); j++) {
                    if (array[j - 1] > array[j]) {
                        //swap elements
                        temp = array[j - 1];
                        array[j - 1] = array[j];
                        array[j] = temp;
                    }

                }
            }

            return array;
        } else {
            return new int[0];
        }
    }

}

